refactor: extract staking into separate module
