+++ pallets/remote-asset-manager/src/calls/proxy.rs
@@ -1,0 +1,4 @@
+// Copyright 2021 ChainSafe Systems
+// SPDX-License-Identifier: LGPL-3.0-only
+
+//! Xcm support for `pallet_proxy` calls
